[BE] refactor: log 클래스 분리 및 패키지 구조 변경 #339
