+++ backend/src/main/java/com/onair/hearit/common/log/config/LogFilterConfig.java
@@ -1,7 +1,7 @@
 package com.onair.hearit.common.log.config;
 
-import com.onair.hearit.common.log.message.MdcLoggingFilter;
-import com.onair.hearit.common.log.message.RequestLoggingFilter;
+import com.onair.hearit.common.log.MdcSetupFilter;
+import com.onair.hearit.common.log.RequestLoggingFilter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -11,8 +11,8 @@ import org.springframework.core.Ordered;
 public class LogFilterConfig {
 
     @Bean
-    public FilterRegistrationBean<MdcLoggingFilter> mdcFilterRegistration(MdcLoggingFilter filter) {
-        FilterRegistrationBean<MdcLoggingFilter> registration = new FilterRegistrationBean<>(filter);
+    public FilterRegistrationBean<MdcSetupFilter> mdcFilterRegistration(MdcSetupFilter filter) {
+        FilterRegistrationBean<MdcSetupFilter> registration = new FilterRegistrationBean<>(filter);
         registration.setOrder(Ordered.HIGHEST_PRECEDENCE);
         registration.addUrlPatterns("/*");
         return registration;
