+++ backend/src/main/java/com/onair/hearit/common/log/dto/ResponseLog.java
@@ -1,4 +1,4 @@
-package com.onair.hearit.common.log.message.dto;
+package com.onair.hearit.common.log.dto;
 
 import java.time.LocalDateTime;
 import org.springframework.http.ResponseEntity;
